+++ pkg/client/connector/install_test.go
@@ -35,7 +35,7 @@ func publishManager(t *testing.T) {
 	// then it builds for the platform indicated by those variables.
 	cmd.Env = []string{
 		"TELEPRESENCE_VERSION=" + version.Version,
-		"TELEPRESENCE_REGISTRY=" + dtest.DockerRegistry(),
+		"TELEPRESENCE_REGISTRY=" + dtest.DockerRegistry(ctx),
 	}
 	includeEnv := []string{"KO_DOCKER_REPO=", "HOME=", "PATH=", "LOGNAME=", "TMPDIR=", "MAKELEVEL="}
 	for _, env := range os.Environ() {
@@ -86,20 +86,21 @@ func removeManager(t *testing.T, kubeconfig, managerNamespace string) {
 }
 
 func TestE2E(t *testing.T) {
-	kubeconfig := dtest.Kubeconfig()
+	ctx := dlog.NewTestContext(t, false)
+	kubeconfig := dtest.Kubeconfig(ctx)
+	registry := dtest.DockerRegistry(ctx)
+
 	testVersion := fmt.Sprintf("v2.0.0-gotest.%d", os.Getpid())
 	namespace := fmt.Sprintf("telepresence-%d", os.Getpid())
 	managerTestNamespace := fmt.Sprintf("ambassador-%d", os.Getpid())
 
-	registry := dtest.DockerRegistry()
 	version.Version = testVersion
 
 	os.Setenv("DTEST_KUBECONFIG", kubeconfig)
 	os.Setenv("KO_DOCKER_REPO", registry)
 	os.Setenv("TELEPRESENCE_REGISTRY", registry)
 
-	dtest.WithMachineLock(func() {
-		ctx := dlog.NewTestContext(t, false)
+	dtest.WithMachineLock(ctx, func(ctx context.Context) {
 		_ = dexec.CommandContext(ctx, "kubectl", "--kubeconfig", kubeconfig, "create", "namespace", namespace).Run()
 		defer func() {
 			_ = dexec.CommandContext(ctx, "kubectl", "--kubeconfig", kubeconfig, "delete", "namespace", managerTestNamespace, "--wait=false").Run()