+++ pkg/client/logging/initcontext.go
@@ -8,18 +8,22 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/client"
 
 	"github.com/sirupsen/logrus"
-	"golang.org/x/crypto/ssh/terminal"
+	"golang.org/x/term"
 
 	"github.com/datawire/dlib/dlog"
 	"github.com/telepresenceio/telepresence/v2/pkg/filelocation"
 )
 
 // IsTerminal returns whether the given file descriptor is a terminal
-var IsTerminal = terminal.IsTerminal
+var IsTerminal = term.IsTerminal
+
+// loggerForTest exposes internals to initcontext_test.go
+var loggerForTest *logrus.Logger
 
 // InitContext sets up standard Telepresence logging for a background process
 func InitContext(ctx context.Context, name string) (context.Context, error) {
-	logger := logrus.StandardLogger()
+	logger := logrus.New()
+	loggerForTest = logger
 	logLevels := client.GetConfig(ctx).LogLevels
 	if name == "daemon" {
 		logger.SetLevel(logLevels.RootDaemon)
@@ -28,9 +32,9 @@ func InitContext(ctx context.Context, name string) (context.Context, error) {
 	}
 
 	if IsTerminal(int(os.Stdout.Fd())) {
-		logger.Formatter = NewFormatter("15:04:05")
+		logger.Formatter = NewFormatter("15:04:05.0000")
 	} else {
-		logger.Formatter = NewFormatter("2006/01/02 15:04:05")
+		logger.Formatter = NewFormatter("2006/01/02 15:04:05.0000")
 		dir, err := filelocation.AppUserLogDir(ctx)
 		if err != nil {
 			return ctx, err