+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
 ### 2.3.0 TBD
-- Bugfix: Using --docker-run no longer fail to mount remote volumes when docker runs as root
+
+- Feature: Telepresence is now installable via brew
 - Change: A TUN-device is used instead of firewall rules for routing outbound connections.
 - Change: Outbound connections now use gRPC instead of ssh, and the traffic-manager no longer has a sshd running.
 - Change: The traffic-agent no longer has a sshd running. Remote volume mounts use sshfs in slave mode, talking directly to sftp.
 - Change: The local DNS now routes the name lookups to intercepted agents or traffic-manager.
-- Feature: Telepresence is now installable via brew
-- Change: The default log-level for the traffic-manager and the root-daemon was changed from "debug' to "info".
+- Change: The default log-level for the traffic-manager and the root-daemon was changed from "debug" to "info".
+- Bugfix: Using --docker-run no longer fail to mount remote volumes when docker runs as root.
+- Bugfix: Fixed a number of race conditions.
 
 ### 2.2.2 (May 17, 2021)
+
 - Feature: Telepresence translates legacy Telepresence commands into viable Telepresence commands.
 - Bugfix: Intercepts will only look for agents that are in the same namespace as the intercept.